adds delete item mutation backend and frontend

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -51,13 +51,20 @@ const Mutations = {
   },
 
   async deleteItem(parent, args, ctx, info) {
-     const where = { id: args.id };
-     // find the item
-     const item = await ctx.db.query.item({ where }, `{id title}`)
-     // check if they own the item, or have permissions
-     // TODO
-     // delete it
-     return ctx.db.mutation.deleteItem({ where }, info);
+    
+    const where = { id: args.id };
+    // find the item
+    const item = await ctx.db.query.item({ where }, `{id title user{ id }}`)
+    // check if they own the item, or have permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(
+      permission => ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if ( !ownsItem && !hasPermissions ) {
+      throw new Error('You don\'t have permission to do that.')
+
+    } 
+    return ctx.db.mutation.deleteItem({ where }, info);
   },
 
   async signup(parent, args, ctx, info) {
